refactor: move server-config state back to core state
- needed for access to locale setting of ICM in core configuration state
- relates to #227
