+++ src/app/core/store/core/server-config/index.ts
