+++ src/app/core/store/core/server-config/server-config.effects.spec.ts
@@ -8,7 +8,6 @@ import { instance, mock, when } from 'ts-mockito';
 
 import { ConfigurationService } from 'ish-core/services/configuration/configuration.service';
 import { CoreStoreModule } from 'ish-core/store/core/core-store.module';
-import { GeneralStoreModule } from 'ish-core/store/general/general-store.module';
 import { makeHttpError } from 'ish-core/utils/dev/api-service-utils';
 import { StoreWithSnapshots, provideStoreSnapshots } from 'ish-core/utils/dev/ngrx-testing';
 import { routerTestNavigationAction } from 'ish-core/utils/dev/routing';
@@ -25,7 +24,7 @@ describe('Server Config Effects', () => {
     when(configurationServiceMock.getServerConfiguration()).thenReturn(of({}));
 
     TestBed.configureTestingModule({
-      imports: [CoreStoreModule.forTesting([], [ServerConfigEffects]), GeneralStoreModule.forTesting('serverConfig')],
+      imports: [CoreStoreModule.forTesting(['serverConfig'], [ServerConfigEffects])],
       providers: [
         provideStoreSnapshots(),
         { provide: ConfigurationService, useFactory: () => instance(configurationServiceMock) },
