+++ src/app/core/store/core/server-config/server-config.effects.ts
