+++ src/app/core/store/core/server-config/server-config.reducer.ts
