+++ src/app/extensions/captcha/facades/captcha.facade.ts
@@ -3,7 +3,7 @@ import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { filter, map, switchMap, switchMapTo } from 'rxjs/operators';
 
-import { getServerConfigParameter } from 'ish-core/store/general/server-config';
+import { getServerConfigParameter } from 'ish-core/store/core/server-config';
 import { whenTruthy } from 'ish-core/utils/operators';
 
 export type CaptchaTopic =
